# tools/sdk-ts/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Notifo API",
    "version": "1.0.0",
    "x-logo": {
      "url": "/logo.svg"
    }
  },
  "servers": [
    {
      "url": "https://localhost:5002"
    }
  ],
  "paths": {
...
    "/api/me/mobilepush": {
      "get": {
        "tags": [
          "MobilePush"
        ],
        "summary": "Returns the mobile push tokens.",
        "operationId": "MobilePush_GetMyToken",
        "responses": {
          "200": {
            "description": "Mobile push tokens returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponseDtoOfMobilePushTokenDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "getMyToken"
      },
      "post": {
        "tags": [
          "MobilePush"
        ],
        "summary": "Register a mobile push token for the current user.",
        "operationId": "MobilePush_PostMyToken",
        "requestBody": {
          "x-name": "request",
          "description": "The request object.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterMobileTokenDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "204": {
            "description": "Mobile push token registered."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "postMyToken"
      }
    },
    "/api/me/mobilepush/{token}": {
      "delete": {
        "tags": [
          "MobilePush"
        ],
        "summary": "Deletes a mobile push token for the current user.",
        "operationId": "MobilePush_DeleteMyToken",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "description": "The token to remove.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Mobile push token removed."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "deleteMyToken"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "statusCode"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "The error message.",
            "nullable": true
          },
          "errorCode": {
            "type": "string",
            "description": "The error code.",
            "nullable": true
          },
          "traceId": {
            "type": "string",
            "description": "The optional trace id.",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "The error type, usually a link.",
            "nullable": true
          },
          "details": {
            "type": "array",
            "description": "Detailed error messages.",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "statusCode": {
            "type": "integer",
            "description": "Status code of the http response.",
            "format": "int32"
          }
        }
      },
...
      "MobileDeviceType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Unknown",
          "Android",
          "iOS"
        ],
        "enum": [
          "Unknown",
          "Android",
          "iOS"
        ]
      },
...
      "ListResponseDtoOfMobilePushTokenDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "description": "The items.",
            "items": {
              "$ref": "#/components/schemas/MobilePushTokenDto2"
            }
          },
          "total": {
            "type": "integer",
            "description": "The total number of items.",
            "format": "int64"
          }
        }
      },
      "MobilePushTokenDto2": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "token",
          "deviceType"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "The device token."
          },
          "deviceType": {
            "description": "The device type.",
            "$ref": "#/components/schemas/MobileDeviceType"
          },
          "deviceIdentifier": {
            "type": "string",
            "description": "A unique identifier for the device.",
            "nullable": true
          }
        }
      },
      "RegisterMobileTokenDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string",
            "description": "The device token."
          },
          "deviceType": {
            "description": "The device type.",
            "$ref": "#/components/schemas/MobileDeviceType"
          },
          "deviceIdentifier": {
            "type": "string",
            "description": "A unique identifier for the device.",
            "nullable": true
          }
        }
      },
...
    }
  }
}